Add bulk get-memory RPC with uint32 length and result pool.
Raise get-memory to 1..65536 (reject address+length wrap), complete via
token-keyed pool + slim MEMORY_RPC_COMPLETE events, and bump the wire
protocol to C64M/2. set-memory stays capped at 1024.

+++ import/c64m/src/control/control_protocol.h
@@ -80,7 +80,7 @@ typedef struct control_args {
     uint32_t id;
     uint32_t timeout_ms;
     uint16_t address;
-    uint16_t length;
+    uint32_t length; /* get-memory: 1..65536; set-memory still 1..1024 */
     uint16_t start_address;
     uint16_t end_address;
     uint16_t run_address;
